# public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: 'API Documentation - Roomac'
    description: ''
    version: 1.0.0
servers:
    -
        url: 'https://it114118-fyp.herokuapp.com'
paths:
    /api/branches/import:
        post:
            summary: "Import branches\n\nImport branches."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: multipart/form-data
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Branch
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                                    description: ''
                            required:
                                - file
    /api/branches/export:
        get:
            summary: "Export branches\n\nExport branches."
            description: ''
            parameters:
                -
                    in: query
                    name: format
                    description: 'Define the export format. Accepted: xlsx, csv, tsv, ods, xls, html. Defaults to xlsx.'
                    example: csv
                    required: false
                    schema:
                        type: string
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                401:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    message: Unauthenticated.
                                properties:
                                    message: { type: string, example: Unauthenticated. }
            tags:
                - Branch
    /api/branches:
        delete:
            summary: "Remove multiple branches\n\nRemove multiple branches."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Branch
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                ids:
                                    type: array
                                    description: 'Array of the branches'' id'
                                    example: '{"ids": ["ST", "TY"]}'
                                    items: { type: string }
                            required:
                                - ids
        get:
            summary: "Retrieve all branches\n\nRetrieve all branches."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                example:
                                    - { id: CW, title_en: 'Chai Wan', title_hk: 柴灣, title_cn: 柴湾, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550859/CW_limxvp.jpg', lat: '22.27184971', lng: '114.23967000', created_at: null, updated_at: null }
                                    - { id: HW, title_en: 'Haking Wong', title_hk: 黃克競, title_cn: 黄克竞, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550860/HW_nhu0jr.jpg', lat: '22.33551569', lng: '114.15235304', created_at: null, updated_at: null }
                                    - { id: KC, title_en: 'Kwai Chung', title_hk: 葵涌, title_cn: 葵涌, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550860/KC_hxxaga.jpg', lat: '22.36187488', lng: '114.12739741', created_at: null, updated_at: null }
                                    - { id: KT, title_en: 'Kwun Tong', title_hk: 觀塘, title_cn: 观塘, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550861/KT_ktbvgg.jpg', lat: '22.31356560', lng: '114.23194114', created_at: null, updated_at: null }
                                    - { id: LWL, title_en: 'Lee Wai Lee', title_hk: 李惠利, title_cn: 李惠利, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550860/LWL_s7uimr.jpg', lat: '22.30620069', lng: '114.25416772', created_at: null, updated_at: null }
                                    - { id: MH, title_en: 'Morrison Hill', title_hk: 摩理臣山, title_cn: 摩理臣山, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550860/MH_tt43bg.jpg', lat: '22.27619053', lng: '114.17792361', created_at: null, updated_at: null }
                                    - { id: ST, title_en: 'Sha Tin', title_hk: 沙田, title_cn: 沙田, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550859/ST_riofal.jpg', lat: '22.39041332', lng: '114.19803821', created_at: null, updated_at: null }
                                    - { id: TM, title_en: 'Tuen Mun', title_hk: 屯門, title_cn: 屯门, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550860/TM_vwyaga.jpg', lat: '22.39311344', lng: '113.96646235', created_at: null, updated_at: null }
                                    - { id: TY, title_en: 'Tsing Yi', title_hk: 青衣, title_cn: 青衣, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1614550859/TY_rli7ox.jpg', lat: '22.34261363', lng: '114.10624981', created_at: null, updated_at: null }
            tags:
                - Branch
            security: []
        post:
            summary: "Add a branch\n\nAdd a branch."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Branch
            security: []
    /api/branches/reset:
        delete:
            summary: "Reset branches\n\nRemove all branches."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Branch
...
    '/api/branches/{branch}/bookings':
        get:
            summary: "Retrieve all branch's bookings\n\nRetrieve all branch's bookings. Example: /api/branches/1/bookings?start=2021-01-24&end=2021-01-30"
            description: ''
            parameters:
                -
                    in: query
                    name: start
                    description: 'query start time in Y-m-d format. Defaults to 2021-01-13.'
                    example: enim
                    required: false
                    schema:
                        type: string
                -
                    in: query
                    name: end
                    description: 'query end time in Y-m-d format. Defaults to 2021-01-15.'
                    example: et
                    required: false
                    schema:
                        type: string
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                401:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    message: Unauthenticated.
                                properties:
                                    message: { type: string, example: Unauthenticated. }
            tags:
                - 'Branch Booking'
            security: []
    '/api/branches/{branch}/settings/active':
        get:
            summary: "Retrieve the active branch's settings\n\nRetrieve the active branch's settings."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    version: null
                                    name: 'Default Settings'
                                    active_at: null
                                    is_active: true
                                    settings: [{ id: CLOSE_TIME, data_type: TIME, default_value: '20:00:00', created_at: null, updated_at: null, value: '20:00:00' }, { id: MIN_CLIENT_UNLOCK, data_type: INTEGER, default_value: '3', created_at: null, updated_at: null, value: '3' }, { id: OPEN_TIME, data_type: TIME, default_value: '08:30:00', created_at: null, updated_at: null, value: '08:30:00' }, { id: RESOURCE_MINUTE_PER_SESSION, data_type: INTEGER, default_value: '30', created_at: null, updated_at: null, value: '30' }, { id: TEST_BOOLEAN_FALSE, data_type: BOOLEAN, default_value: '0', created_at: null, updated_at: null, value: '0' }, { id: TEST_BOOLEAN_TRUE, data_type: BOOLEAN, default_value: '1', created_at: null, updated_at: null, value: '1' }, { id: TEST_VARCHAR, data_type: VARCHAR, default_value: 'Test String', created_at: null, updated_at: null, value: 'Test String' }, { id: TIME_IN_ADVANCE, data_type: TIME, default_value: '24:00:00', created_at: null, updated_at: null, value: '24:00:00' }]
                                properties:
                                    version: { type: 'null', example: null }
                                    name: { type: string, example: 'Default Settings' }
                                    active_at: { type: 'null', example: null }
                                    is_active: { type: boolean, example: true }
                                    settings: { type: array, example: [{ id: CLOSE_TIME, data_type: TIME, default_value: '20:00:00', created_at: null, updated_at: null, value: '20:00:00' }, { id: MIN_CLIENT_UNLOCK, data_type: INTEGER, default_value: '3', created_at: null, updated_at: null, value: '3' }, { id: OPEN_TIME, data_type: TIME, default_value: '08:30:00', created_at: null, updated_at: null, value: '08:30:00' }, { id: RESOURCE_MINUTE_PER_SESSION, data_type: INTEGER, default_value: '30', created_at: null, updated_at: null, value: '30' }, { id: TEST_BOOLEAN_FALSE, data_type: BOOLEAN, default_value: '0', created_at: null, updated_at: null, value: '0' }, { id: TEST_BOOLEAN_TRUE, data_type: BOOLEAN, default_value: '1', created_at: null, updated_at: null, value: '1' }, { id: TEST_VARCHAR, data_type: VARCHAR, default_value: 'Test String', created_at: null, updated_at: null, value: 'Test String' }, { id: TIME_IN_ADVANCE, data_type: TIME, default_value: '24:00:00', created_at: null, updated_at: null, value: '24:00:00' }], items: { type: object } }
            tags:
                - 'Branch Setting'
            security: []
        parameters:
            -
                in: path
                name: branch
                description: 'Optional parameter. string required The ID of the branch.'
                required: true
                schema:
                    type: string
                examples:
                    omitted:
                        summary: 'When the value is omitted'
                        value: ''
                    present:
                        summary: 'When the value is present'
                        value: ST
    '/api/branches/{branch}/settings':
        get:
            summary: "Retrieve all branch's settings versions\n\nRetrieve all branch's settings versions."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    active_version: null
                                    versions: [{ version: null, name: 'Default Settings' }]
                                properties:
                                    active_version: { type: 'null', example: null }
                                    versions: { type: array, example: [{ version: null, name: 'Default Settings' }], items: { type: object } }
            tags:
                - 'Branch Setting'
            security: []
        post:
            summary: "Add a new version of branch's settings\n\nAdd a new version of branch's settings"
            description: ''
            parameters:
                -
                    in: query
                    name: name
                    description: 'The name of the setting. Defaults to "New Settings".'
                    example: 'My Special Settings'
                    required: false
                    schema:
                        type: string
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - 'Branch Setting'
            security: []
        parameters:
            -
                in: path
                name: branch
                description: 'Optional parameter. string required The ID of the branch.'
                required: true
                schema:
                    type: string
                examples:
                    omitted:
                        summary: 'When the value is omitted'
                        value: ''
                    present:
                        summary: 'When the value is present'
                        value: ST
...
    /api/categories/import:
        post:
            summary: "Import categories\n\nImport categories."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: multipart/form-data
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Category
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                                    description: ''
                            required:
                                - file
    /api/categories/export:
        get:
            summary: "Export categories\n\nExport categories."
            description: ''
            parameters:
                -
                    in: query
                    name: format
                    description: 'Define the export format. Accepted: xlsx, csv, tsv, ods, xls, html. Defaults to xlsx.'
                    example: csv
                    required: false
                    schema:
                        type: string
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Category
    /api/categories:
        delete:
            summary: "Remove multiple categories\n\nRemove multiple categories."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Category
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                ids:
                                    type: array
                                    description: 'Array of the categories'' id'
                                    example: '{"ids": [1, 2]}'
                                    items: { type: string }
                            required:
                                - ids
        get:
            summary: "Retrieve all categories\n\nRetrieve all categories."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                example:
                                    - { id: 1, title_en: Classroom, title_hk: 課室, title_cn: 教室, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1610342834/riak0mox4pqzxesenegs.jpg', created_at: null, updated_at: null }
                                    - { id: 2, title_en: Library, title_hk: 圖書館, title_cn: 图书馆, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1610343016/ca8zmlcwcbcspgw6sked.jpg', created_at: null, updated_at: null }
                                    - { id: 3, title_en: 'Computer Room', title_hk: 電腦房, title_cn: 电脑房, image_url: 'https://res.cloudinary.com/hkzbjzedn/image/upload/v1611592893/o72p9styjrmdhbrf77zw.jpg', created_at: null, updated_at: null }
            tags:
                - Category
            security: []
        post:
            summary: "Add a category\n\nAdd a category."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Category
    /api/categories/reset:
        delete:
            summary: "Reset categories\n\nRemove all categories."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Category
...
    /api/login:
        post:
            summary: 'Authenticate a user'
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: '200, success'
                    content:
                        text/plain:
                            schema:
                                type: string
                                example: 1|sNt8wF0Zh4oGJ20O22gns0K4bI2HJfkqNZWiKoEX
            tags:
                - Login
            security: []
    /api/logout:
        post:
            summary: 'Logout a user'
            description: ''
            parameters:
                -
                    in: query
                    name: global
                    description: 'When this value is set to true all of the tokens issued to the user will be revoked. Defaults to false.'
                    example: null
                    required: false
                    schema:
                        type: string
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Login
    /api/programs/import:
        post:
            summary: "Import programs\n\nImport programs."
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: multipart/form-data
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Program
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                file:
                                    type: string
                                    format: binary
                                    description: ''
                            required:
                                - file
...